requirements.txt
#Modules Needed
python-valve, asyncio, mariadb, discord, sqlite3, folium, python-a2s, lovely_logger
mariadb~=1.0.11
discord~=1.7.3